read columns values by rowid and convert to write columns.
