+++ be/src/storage/rowset_update_state.cpp
@@ -184,9 +184,6 @@ static void plan_read_by_rssid(const vector<uint64_t>& rowids, size_t* num_defau
     }
 }
 
-static void gen_column_by_src_column_and_idxes(const vectorized::Column& src, const vector<uint32_t>& idxes,
-                                               std::unique_ptr<vectorized::Column>* dest) {}
-
 Status RowsetUpdateState::apply(Tablet* tablet, Rowset* rowset, uint32_t rowset_id, const PrimaryIndex& index) {
     const auto& rowset_meta_pb = rowset->rowset_meta()->get_meta_pb();
     if (!rowset_meta_pb.has_txn_meta()) {
@@ -208,6 +205,13 @@ Status RowsetUpdateState::apply(Tablet* tablet, Rowset* rowset, uint32_t rowset_
     }
     vector<std::unique_ptr<vectorized::Column>> read_columns(read_column_ids.size());
     vector<std::unique_ptr<vectorized::Column>> write_columns(read_column_ids.size());
+    for (auto i = 0; read_column_ids.size(); i++) {
+        const auto read_column_id = read_column_ids[i];
+        auto tablet_column = tschema.column(read_column_id);
+        auto column = ChunkHelper::column_from_field_type(tablet_column.type(), tablet_column.is_nullable());
+        read_columns[i] = column->clone_empty();
+        write_columns[i] = column->clone_empty();
+    }
     size_t num_segments = rowset->num_segments();
     DCHECK(num_segments == _upserts.size());
     vector<std::pair<string, string>> rewrite_files;
@@ -220,8 +224,8 @@ Status RowsetUpdateState::apply(Tablet* tablet, Rowset* rowset, uint32_t rowset_
         auto& pks = *_upserts[i];
         int64_t t_start = MonotonicMillis();
         std::vector<uint64_t> rowids(pks.size());
-        // TODO: get rowid by pk column using primary index
         int64_t t_get_rowids = MonotonicMillis();
+        index.get(pks, &rowids);
         size_t num_default = 0;
         std::map<uint32_t, std::vector<uint32_t>> rowids_by_rssid;
         vector<uint32_t> idxes;
@@ -231,7 +235,7 @@ Status RowsetUpdateState::apply(Tablet* tablet, Rowset* rowset, uint32_t rowset_
         RETURN_IF_ERROR(tablet->updates()->_get_column_values(read_column_ids, num_default > 0, rowids_by_rssid,
                                                               &read_columns));
         for (size_t col_idx = 0; col_idx < read_column_ids.size(); i++) {
-            gen_column_by_src_column_and_idxes(*read_columns[col_idx], idxes, &write_columns[col_idx]);
+            write_columns[col_idx]->append_selective(*read_columns[col_idx], idxes.data(), 0, idxes.size());
         }
         int64_t t_get_column_values = MonotonicMillis();
         auto src_path = BetaRowset::segment_file_path(tablet->schema_hash_path(), rowset->rowset_id(), i);
